% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PheGWASData.R
\docType{data}
\name{hdl}
\alias{hdl}
\title{HDL GWAS summary dataset (made available from Lipid consortium data files)}
\format{
A dataframe with simulated GWAS results (simulated using plink 1.9)
\itemize{
  \item CHR is the chromosome number
  \item BP is the position of the chromosome
  \item rsid The SNP ID associated with that position and chromsome
  \item HDL is the p_value of the trait
  \item gene is the gene associated with rsid
}
}
\usage{
data(hdl)
}
\value{
A data frame
}
\description{
The dataset 'hdl' is a data.frame object with the summary results of a GWAS analysis.
}
\keyword{datasets}
