% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PheGWAS.R
\name{landscapetable}
\alias{landscapetable}
\title{3-D association table for many phenotypes}
\usage{
landscapetable(d, sliceval = 8, chromosome = 1, bpdivision = 1e+06)
}
\arguments{
\item{d}{DataFrame output from processphegwas}

\item{sliceval}{Integer to indicate value of -log10(p) to do the sectionalcut. Usually value > -log10 6 is considered to be significant}

\item{chromosome}{Integer to indicate the chromosome number thats interested, If not given entire chromosome is given}

\item{bpdivision}{Integer value to indicate the base pair divisions. By default it is 100,000 bp. This vis only applciable on a single chromosome view. For entire
chromosoem view the max peak is selected}
}
\description{
3-D association table for many phenotypes
}
\examples{
\dontrun{
# table of chromosome number 16, with bp division on 1,000,000 (default is 1,000,000)
x <- list(hdl,ldl,trig,tchol)
phenos <- c("HDL","LDL","TRIGS","TOTALCHOLESTROL")
y <- processphegwas(x, phenos)
landscapetable(y,sliceval = 6,chromosome = 16,bpdivision = 1000000)
}
}
\author{
George Gittu
}
