% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PheGWAS.R
\name{processphegwas}
\alias{processphegwas}
\title{Prepare the dataframe to pass to landscape function}
\usage{
processphegwas(x, phenos)
}
\arguments{
\item{x}{List of dataframes that need to do PheGWAS on. Arrange the dataframe in the order how the the phenotypes should align in y axis}

\item{phenos}{a vector of phenotypes that you are passing for in the summary statistics file. It should be in the order that you pass the List.}
}
\value{
A processed dataframe to pass to PheGWAS function
}
\description{
Prepare the dataframe to pass to landscape function
}
\details{
Make sure there are no duplicate rsid's in any of the dataframe, If there aremake sure to resolve it before passing it to this function.
}
\examples{
\dontrun{
x <- list(hdl,ldl,trig,tchol)
phenos <- c("HDL","LDL","TRIGS","TOTALCHOLESTROL")
## y is ready to be passed to function landscape
y <- processphegwas(x, phenos)
}
}
\author{
George Gittu
}
