% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PheGWAS.R
\name{landscape}
\alias{landscape}
\title{Interactive 3-D association landscape for many phenotypes}
\usage{
landscape(d, sliceval = 0, chromosome = FALSE, bpdivision = 1e+06,
  upperlimit = NULL, geneview = FALSE)
}
\arguments{
\item{d}{DataFrame output from processphegwas}

\item{sliceval}{Integer to indicate value of -log10(p) to do the sectionalcut. Usually value > -log10 6 is considered to be significant}

\item{chromosome}{Integer to indicate the chromosome number thats interested, If not given entire chromosome is given}

\item{bpdivision}{Integer value to indicate the base pair divisions. By default it is 100,000 bp. This vis only applciable on a single chromosome view. For entire
chromosoem view the max peak is selected}
}
\description{
Interactive 3-D association landscape for many phenotypes
}
\examples{
y <- processphegwas(x)
## pass the dataframe from the processphegwas

# 3D landscape visualization of all the phenotypes across the base pair positions
landscape(y)

# 3D landscape visualization of all the phenotypes across the base pair positions, showing only the crusts above a certain threshold
landscape(y,sliceval = 6)

# 3D landscape visualization of chromosome number 16
landscape(x,sliceval = 6,chromosome = 16)

# 3D landscape visualization of chromosome number 16, with bp division on 1,000,000 (default is 100,000)
landscape(x,sliceval = 6,chromosome = 16,bpdivision = 1,000,000)
}
\author{
George Gittuuuuu
}
