% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PheGWAS.R
\name{processphegwas}
\alias{processphegwas}
\title{Prepare the dataframe to pass to landscape function}
\usage{
processphegwas(x)
}
\arguments{
\item{x}{List of dataframes that need to do PheGWAS on. Arrange the dataframe in the order how the the phenotypes should align in y axis}
}
\value{
A processed dataframe to pass to PheGWAS function
}
\description{
Prepare the dataframe to pass to landscape function
}
\details{
Make sure there are no duplicate rsid's in any of the dataframe, If there aremake sure to resolve it before passing it to this function.
}
\examples{
x <- list(bmimen,bmiwomen)

## y is ready to be passed to function landscape
y <- processphegwas(x)
}
\author{
George Gittu
}
